+++ packages/polaris-viz/CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Changed
+
+- Removed comparison color from labels in `<SimpleBarChart />`.
+- Updated `<TrendIndicator />` negative color to neutral grey.
 
 ## [13.2.0] - 2024-05-27